Fixing compilation warnings for r19354
TEST=none
BUG=none

TBR=eroman

Review URL: http://codereview.chromium.org/149074

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@19356 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/socket/socks_client_socket.cc
@@ -299,7 +299,7 @@ int SOCKSClientSocket::DoHandshakeWriteComplete(int result) {
   if (bytes_sent_ == buffer_.size()) {
     next_state_ = STATE_HANDSHAKE_READ;
     buffer_.clear();
-  } else if (bytes_sent_ < buffer_.size()) {
+  } else if (bytes_sent_ < static_cast<int>(buffer_.size())) {
     next_state_ = STATE_HANDSHAKE_WRITE;
   } else {
     return ERR_UNEXPECTED;
